move change_dict_keys_case
